Add delete action
Instead of moving the queues which is prone to errors and to
leave broken queues, we can simply delete the queues wich are
empty and they will be recreated in less than a second and
the proper balincing done to them.

This allows us to be blazing fast with minimal impact and without
having to doo too much.

Wont remove the code from the move part yet, that will come in a
different PR